added states api call + async view
