+++ foia_integration/foia/templates/base_generic.html
@@ -52,4 +52,6 @@
         {% block content %}{% endblock %}
     </main> 
 </body>
+{% block inline_scripts %}
+{% endblock %}
 </html>
