+++ foia_integration/foia/urls.py
@@ -6,5 +6,6 @@ urlpatterns = [
     path("template-builder", views.template_render, name="template"),
     path("foia-request", views.file_requests, name="foia-request"),
     # API Calls
+    re_path(r"^api/current-user/states$", api.get_states),
     re_path(r"^api/current-user/template/(?P<state_abbr>[A-Za-z]{2})$", api.get_template)
 ]
