+++ foia_integration/precompiles/svelte/foia-request/src/main.js
@@ -1,13 +1,7 @@
 import App from './App.svelte';
 
 const app = new App({
-	target: document.getElementById("foia-request-items"),
-	props: {
-		states: [
-			{abbr: "AK", name: "Alaska"},
-			{abbr: "CA", name: "California"}
-		]
-	}
+	target: document.getElementById("foia-request-items")
 });
 
 export default app;
